refactor: moved logic for image resize to utils and reduce repetition
