+++ filter-definition.txt
@@ -1,0 +1,1 @@
+path .env`
