+++ src/utils/resize.js
@@ -16,3 +16,27 @@ export const resizeFile = (file) =>
       300
     );
   });
+
+export const alterImageDimensions = async (
+  file,
+  setFileName,
+  setFileUpload
+) => {
+  try {
+    const image = await resizeFile(file);
+    // Create a new File object with the resized image data
+    const resizedFile = new File([image], file.name, { type: file.type });
+    if (resizedFile.name.length > 18) {
+      let fileExt = resizedFile.name.match(/\.[0-9a-z]+$/i)[0];
+      let truncFileName = resizedFile.name.slice(0, 8);
+      const newFileName = `${truncFileName}${fileExt}`;
+      setFileName(newFileName);
+    } else {
+      setFileName(resizedFile.name);
+    }
+
+    setFileUpload(resizedFile);
+  } catch (err) {
+    console.log(err);
+  }
+};
